refactor (Player) : 패키지 변경 및 클래스 이름 변경
- 클래스 이름에서 필요 없는 `Name` 제거
- 한 패키지에 많은 클래스가 생기는 것을 막기 위해 새로운 패키지 생성
