+++ src/main/java/ladder/domain/player/Player.java
@@ -1,12 +1,12 @@
-package ladder.domain;
+package ladder.domain.player;
 
-public class PlayerName {
+public class Player {
 
     private static final int MAX_LENGTH = 5;
 
     private final String name;
 
-    public PlayerName(String name) {
+    public Player(String name) {
         validate(name);
         this.name = name;
     }
