+++ src/test/java/ladder/domain/player/PlayerTest.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.player;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -9,14 +9,14 @@ import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
 
-class PlayerNameTest {
+class PlayerTest {
 
     @DisplayName("이름은 5글자를 넘을 수 없다")
     @ParameterizedTest
     @CsvSource({"steve!", "123 45", "123456"})
     void validateTest_WhenLengthIsOver5(String name) {
 
-        assertThatThrownBy(() -> new PlayerName(name))
+        assertThatThrownBy(() -> new Player(name))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("이름은 5글자를 넘을 수 없습니다.");
     }
@@ -27,7 +27,7 @@ class PlayerNameTest {
     @ValueSource(strings = " ")
     void validateTest_WhenNameIsEmpty(String name) {
 
-        assertThatThrownBy(() -> new PlayerName(name))
+        assertThatThrownBy(() -> new Player(name))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("이름은 한글자 이상이어야 합니다.");
     }
@@ -36,8 +36,8 @@ class PlayerNameTest {
     @ParameterizedTest
     @ValueSource(strings = {"steve", "lucy", "bri"})
     void validateTest_WhenNameSizeIsInRange(String name) {
-        PlayerName playerName = new PlayerName(name);
+        Player player = new Player(name);
 
-        assertThat(playerName.getName()).isEqualTo(name);
+        assertThat(player.getName()).isEqualTo(name);
     }
 }
